+++ j/src/org/armedbear/lisp/Nil.java
@@ -2,7 +2,7 @@
  * Nil.java
  *
  * Copyright (C) 2002-2003 Peter Graves
- * $Id: Nil.java,v 1.9 2003-03-09 03:41:03 piso Exp $
+ * $Id: Nil.java,v 1.10 2003-03-18 03:45:49 piso Exp $
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -76,7 +76,7 @@ public final class Nil extends LispObject implements SequenceType
         return 0;
     }
 
-    public LispObject elt(long index) throws LispError
+    public LispObject elt(int index) throws LispError
     {
         throw new TypeError("ELT: invalid index " + index + " for " + this);
     }
